feature: add NextGenTestGatewaySubscriptionModule and register in service provider (#149)
