+++ src/NextGen/Gateways/NextGenTestGateway/NextGenTestGateway.php
@@ -2,17 +2,12 @@
 
 namespace Give\NextGen\Gateways\NextGenTestGateway;
 
-use Exception;
 use Give\Donations\Models\Donation;
 use Give\Framework\EnqueueScript;
-use Give\Framework\PaymentGateways\Commands\GatewayCommand;
 use Give\Framework\PaymentGateways\Commands\PaymentComplete;
-use Give\Framework\PaymentGateways\Commands\SubscriptionComplete;
 use Give\Framework\PaymentGateways\Contracts\NextGenPaymentGatewayInterface;
 use Give\Framework\PaymentGateways\PaymentGateway;
 use Give\NextGen\Framework\PaymentGateways\Traits\HandleHttpResponses;
-use Give\Subscriptions\Models\Subscription;
-use Give\Subscriptions\ValueObjects\SubscriptionStatus;
 
 /**
  * @since 0.1.0
@@ -88,34 +83,6 @@ class NextGenTestGateway extends PaymentGateway implements NextGenPaymentGateway
         return new PaymentComplete($transactionId);
     }
 
-    /**
-     * @inheritDoc
-     *
-     * @unreleased
-     */
-    public function createSubscription(
-        Donation $donation,
-        Subscription $subscription,
-        $gatewayData = null
-    ): GatewayCommand {
-        return new SubscriptionComplete(
-            "test-gateway-transaction-id-$donation->id",
-            "test-gateway-subscription-id-$subscription->id"
-        );
-    }
-
-    /**
-     * @unreleased
-     *
-     * @inheritDoc
-     * @throws Exception
-     */
-    public function cancelSubscription(Subscription $subscription)
-    {
-        $subscription->status = SubscriptionStatus::CANCELLED();
-        $subscription->save();
-    }
-
 
     /**
      * @inerhitDoc
@@ -125,6 +92,9 @@ class NextGenTestGateway extends PaymentGateway implements NextGenPaymentGateway
         // TODO: Implement refundDonation() method.
     }
 
+    /**
+     * @since 0.1.0
+     */
     public function formSettings(int $formId): array
     {
         return [];
