+++ src/NextGen/Gateways/NextGenTestGateway/NextGenTestGatewaySubscriptionModule.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace Give\NextGen\Gateways\NextGenTestGateway;
+
+use Exception;
+use Give\Donations\Models\Donation;
+use Give\Framework\PaymentGateways\Commands\SubscriptionComplete;
+use Give\Framework\PaymentGateways\Contracts\Subscription\SubscriptionAmountEditable;
+use Give\Framework\PaymentGateways\SubscriptionModule;
+use Give\Framework\Support\ValueObjects\Money;
+use Give\Subscriptions\Models\Subscription;
+use Give\Subscriptions\ValueObjects\SubscriptionStatus;
+
+class NextGenTestGatewaySubscriptionModule extends SubscriptionModule implements SubscriptionAmountEditable
+{
+    /**
+     * @unreleased
+     */
+    public function createSubscription(
+        Donation $donation,
+        Subscription $subscription,
+        $gatewayData = null
+    ): SubscriptionComplete {
+        return new SubscriptionComplete(
+            "test-gateway-transaction-id-$donation->id",
+            "test-gateway-subscription-id-$subscription->id"
+        );
+    }
+
+
+    /**
+     * @unreleased
+     *
+     * @throws Exception
+     */
+    public function cancelSubscription(Subscription $subscription)
+    {
+        $subscription->status = SubscriptionStatus::CANCELLED();
+        $subscription->save();
+    }
+
+    /**
+     * @unreleased
+     * 
+     * @throws Exception
+     */
+    public function updateSubscriptionAmount(Subscription $subscription, Money $newRenewalAmount)
+    {
+        $subscription->amount = $newRenewalAmount;
+        $subscription->save();
+    }
+}
